+++ minivideo/src/decoder/h264/h264_slice_struct.h
@@ -125,12 +125,12 @@ typedef struct drpm_t
 //! CABAC decoding context
 typedef struct CabacContext_t
 {
-    int pStateIdx[460]; //!< Corresponds to a probability state index
-    int valMPS[460];    //!< Corresponds to the value of the most probable symbol
+    uint8_t pStateIdx[460]; //!< Corresponds to a probability state index
+    uint8_t valMPS[460];    //!< Corresponds to the value of the most probable symbol
 
     // The status of the arithmetic decoding engine is represented by the variables codIRange and codIOffset.
-    int codIRange;
-    int codIOffset;
+    uint16_t codIRange;
+    uint16_t codIOffset;
 } CabacContext_t;
 
 /*!
